+++ R/plotsNtables.R
@@ -242,7 +242,8 @@ summaryBarPlot <- function(bt_summary, measures = NULL, type = c("ggplot2", "sim
 backtestBoxPlot <- function(bt, measure = "Sharpe ratio", type = c("ggplot2", "simple"), ...) {
   # extract correct performance measure
   res_list_table <- backtestTable(bt)
-  idx <- grep(measure, names(res_list_table), ignore.case = TRUE)
+  # idx <- grep(measure, names(res_list_table), ignore.case = TRUE)  # it does not work when "measure" contains brackets
+  idx <- which(measure == names(res_list_table))
   if (length(idx)!=1) stop(measure, "does not match a single performance measure")
   res_table <- res_list_table[[idx]]